перенес генерацию дней в utils, рефакторинг рендер-метода
