+++ src/utils.js
@@ -1,4 +1,4 @@
-export const renderElement = (container, element, place) => {
+export const render = (container, element, place) => {
   switch (place) {
     case RenderPosition.BEFOREBEGIN:
       container.before(element);
@@ -60,5 +60,34 @@ export const createDateInAttributeFormat = (date) => {
   const month = `0${date.getMonth() + 1}`;
   const day = date.getDate();
   return `${year}-${month}-${day}`;
+};
+
+export const generateDays = (events) => {
+  const eventsDates = events.map((event) => {
+    return event.dateStart;
+  });
+
+  eventsDates.sort((a, b) => a - b);
+
+  let days = new Set();
+
+  eventsDates.forEach((date) => {
+    days.add(date);
+  });
+
+  days = Array.from(days).map((day) => {
+    const eventsInDay = events.filter((event) => {
+      return event.dateStart === day;
+    });
+    eventsInDay.sort((a, b) => {
+      return Date.parse(a.dateTimeStart) - Date.parse(b.dateTimeStart);
+    });
+    return {
+      day,
+      events: eventsInDay,
+    };
+  });
+
+  return days;
 };
